% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recplot.R
\name{RecruitPlotEasy}
\alias{RecruitPlotEasy}
\title{Run RecruitPlotEasy}
\usage{
RecruitPlotEasy()
}
\description{
This function controls all of the behavior of RecruitPlotEasy, from preparing the
python environment to launching the GUI that you're going to interact with.
}
\details{
It has no parameters
}
